NFC: move the IMT builder to a separate file
