+++ src/mono/mono/metadata/CMakeLists.txt
@@ -82,6 +82,7 @@ set(metadata_common_sources
     icall-eventpipe.c
     image.c
     image-internals.h
+    imt-builder.c
     webcil-loader.h
     webcil-loader.c
     jit-info.h
